Deploy stake (#73)
* Massive refactoring of deploy

* Improvements and deploy for StakeConfigurator and stake tokens 

* More stake events

+++ contracts/protocol/stake/interfaces/IStakeToken.sol
@@ -35,9 +35,9 @@ interface IStakeToken is IDerivedToken {
 interface IManagedStakeToken is IEmergencyAccess {
   function setRedeemable(bool redeemable) external;
 
-  function getMaxSlashablePercentage() external view returns (uint256);
+  function getMaxSlashablePercentage() external view returns (uint16);
 
-  function setMaxSlashablePercentage(uint256 percentage) external;
+  function setMaxSlashablePercentage(uint16 percentage) external;
 
   function setCooldown(uint32 cooldownPeriod, uint32 unstakePeriod) external;
 }
